+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Features
 
+-   [`4d0ce60`](https://github.com/stdlib-js/stdlib/commit/4d0ce60011997633aac407dbba435da1763a2550) - add float16 dtype support to `array/convert-same` [(#14126)](https://github.com/stdlib-js/stdlib/pull/14126)
 -   [`823ccaa`](https://github.com/stdlib-js/stdlib/commit/823ccaae2a34ef2fc8d0795f885b0503394ddae8) - add float16 dtype support to `array/convert` [(#14125)](https://github.com/stdlib-js/stdlib/pull/14125)
 -   [`f95a2e1`](https://github.com/stdlib-js/stdlib/commit/f95a2e1e1f7fb5bd69f9d403cc8f9b40a7b1defb) - add float16 dtype support to `array/nans` [(#14127)](https://github.com/stdlib-js/stdlib/pull/14127)
 -   [`28c1b87`](https://github.com/stdlib-js/stdlib/commit/28c1b877001ac4db853b9a958d9ec33bd1bd8288) - add float16 dtype support to `array/one-to-like` [(#14148)](https://github.com/stdlib-js/stdlib/pull/14148)
@@ -85,6 +86,7 @@ This release closes the following issue:
 
 <details>
 
+-   [`4d0ce60`](https://github.com/stdlib-js/stdlib/commit/4d0ce60011997633aac407dbba435da1763a2550) - **feat:** add float16 dtype support to `array/convert-same` [(#14126)](https://github.com/stdlib-js/stdlib/pull/14126) _(by Gururaj Gurram)_
 -   [`823ccaa`](https://github.com/stdlib-js/stdlib/commit/823ccaae2a34ef2fc8d0795f885b0503394ddae8) - **feat:** add float16 dtype support to `array/convert` [(#14125)](https://github.com/stdlib-js/stdlib/pull/14125) _(by Gururaj Gurram)_
 -   [`f95a2e1`](https://github.com/stdlib-js/stdlib/commit/f95a2e1e1f7fb5bd69f9d403cc8f9b40a7b1defb) - **feat:** add float16 dtype support to `array/nans` [(#14127)](https://github.com/stdlib-js/stdlib/pull/14127) _(by Gururaj Gurram, Athan Reines)_
 -   [`28c1b87`](https://github.com/stdlib-js/stdlib/commit/28c1b877001ac4db853b9a958d9ec33bd1bd8288) - **feat:** add float16 dtype support to `array/one-to-like` [(#14148)](https://github.com/stdlib-js/stdlib/pull/14148) _(by Gururaj Gurram)_